Fix a bug in ParseSpannerType when parsing an ARRAY data type and fix related test code.
With this schema, `gcsb load` causes a panic error, e.g. "panic: unknown spanner type 'STRING(256)'"
CREATE TABLE item_list(
  item_list_id STRING(256),
  item_list1 ARRAY<STRING(256)>,
  item_list2 ARRAY<STRING(MAX)>
) PRIMARY KEY(item_list_id)

This is because the current code doesn't take account of the data length of the inner STRING element. Actually, using STRING data type without specifying the data length is not allowed in DDL and you can't create a table with a data type of `STRING` or ARRAY<STRING>.

This change is to fix the code and related test code.